requirements.txt
# Requirements pour Ayanna ERP
PyQt6>=6.5.0
SQLAlchemy>=2.0.0
bcrypt>=4.0.0
reportlab>=4.0.0
Pillow>=10.0.0
python-dotenv>=1.0.0
